[level 1] Title: 부족한 금액 계산하기, Time: 0.11 ms, Memory: 74.1 MB -BaekjoonHub

+++ 프로그래머스/1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.java
@@ -5,7 +5,7 @@ class Solution {
         
         // count만큼 계산
         for(int i=2; i<=count; i++) {
-            sum += (long)price * i;
+            sum += price * i;
         }
 
         // 모자란 금액
